# script/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Il microservizio espone i metodi per il DMS.",
        "version": "1.0.0",
        "title": "dms-misura-service"
    },
    "host": "localhost:8090",
    "basePath": "/dms-misura-service",
...
    "paths": {
...
        "/documents": {
            "get": {
                "tags": [ "document-controller" ],
                "summary": "list",
                "operationId": "listUsingGET",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "linea_in",
                    "in": "query",
                    "description": "the list of 'linea' to include the research",
                    "required": false,
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "collectionFormat": "multi"
                }, {
                    "name": "sort",
                    "in": "query",
                    "description": "sort field, it indicates the id to perform the sorting",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "direction",
                    "in": "query",
                    "description": "sort direction, must be 'asc' or 'desc' (default = 'asc')",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "limit",
                    "in": "query",
                    "description": "limit the number of records given as result, must be an integer greater than 0 and less than 100",
                    "required": false,
                    "type": "integer",
                    "format": "int64"
                }, {
                    "name": "offset",
                    "in": "query",
                    "description": "the offset from where start giving the result, must be an integer greater than 0",
                    "required": false,
                    "type": "integer",
                    "format": "int64"
                }, {
                    "name": "name",
                    "in": "query",
                    "description": "name of the document",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "status",
                    "in": "query",
                    "description": "document status, can be one of 'active', 'inactive'",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "data_documento_gt",
                    "in": "query",
                    "description": "start of date range search",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "data_documento_lt",
                    "in": "query",
                    "description": "end of date range search",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "linea",
                    "in": "query",
                    "description": "linea ",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "codice_centro",
                    "in": "query",
                    "description": "codice centro ",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "codice_distretto",
                    "in": "query",
                    "description": "codice distretto",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "codice_area_tecnica",
                    "in": "query",
                    "description": "codice area tecnica",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "codice_polo",
                    "in": "query",
                    "description": "codice polo",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "ragione_sociale",
                    "in": "query",
                    "description": "rargione sociale",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "ragione_sociale_like",
                    "in": "query",
                    "description": "substring of rargione sociale",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "ubicazione",
                    "in": "query",
                    "description": "ubicazione",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "ubicazione_like",
                    "in": "query",
                    "description": "substring of ubicazione",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "provincia",
                    "in": "query",
                    "description": "provincia",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "comune",
                    "in": "query",
                    "description": "comune",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "regione",
                    "in": "query",
                    "description": "regione",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "codice_aop",
                    "in": "query",
                    "description": "codice aop",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "codice_ateco",
                    "in": "query",
                    "description": "codice ateco",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "data_entrata_esercizio_gt",
                    "in": "query",
                    "description": "start of data entrata esercizio",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "data_entrata_esercizio_lt",
                    "in": "query",
                    "description": "end of data entrata esercizio",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "folder",
                    "in": "query",
                    "description": "folder",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "subfolder",
                    "in": "query",
                    "description": "subfolder",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "remi_terzo",
                    "in": "query",
                    "description": "remi terzo",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "reti_trasporto",
                    "in": "query",
                    "description": "reti trasporto",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "remi",
                    "in": "query",
                    "description": "the remi",
                    "required": false,
                    "type": "string"
                } ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Document"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [ "document-controller" ],
                "summary": "post",
                "operationId": "postUsingPOST",
                "consumes": [ "multipart/form-data" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "document",
                    "in": "formData",
                    "description": "document",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "file",
                    "in": "formData",
                    "description": "file",
                    "required": true,
                    "type": "file"
                } ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Document"
                        }
                    }
                }
            }
        },
        "/documents/{document_id}": {
            "get": {
                "tags": [ "document-controller" ],
                "summary": "get",
                "operationId": "getUsingGET",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "document_id",
                    "in": "path",
                    "description": "document_id",
                    "required": true,
                    "type": "string"
                } ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Document"
                        }
                    }
                }
            },
            "put": {
                "tags": [ "document-controller" ],
                "summary": "put",
                "operationId": "putUsingPUT",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "document_id",
                    "in": "path",
                    "description": "document_id",
                    "required": true,
                    "type": "string"
                }, {
                    "in": "body",
                    "name": "params",
                    "description": "params",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/DocumentUpdate"
                    }
                } ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Document"
                        }
                    }
                }
            },
            "delete": {
                "tags": [ "document-controller" ],
                "summary": "delete",
                "operationId": "deleteUsingDELETE",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "document_id",
                    "in": "path",
                    "description": "document_id",
                    "required": true,
                    "type": "string"
                } ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/documents/{document_id}/attachments": {
            "get": {
                "tags": [ "document-controller" ],
                "summary": "listAttachment",
                "operationId": "listAttachmentUsingGET",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "sort",
                    "in": "query",
                    "description": "sort field, it indicates the id to perform the sorting",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "direction",
                    "in": "query",
                    "description": "sort direction, must be 'asc' or 'desc' (default = 'asc')",
                    "required": false,
                    "type": "string"
                }, {
                    "name": "limit",
                    "in": "query",
                    "description": "limit the number of records given as result, must be an integer greater than 0 and less than 100",
                    "required": false,
                    "type": "integer",
                    "format": "int64"
                }, {
                    "name": "offset",
                    "in": "query",
                    "description": "the offset from where start giving the result, must be an integer greater than 0",
                    "required": false,
                    "type": "integer",
                    "format": "int64"
                } ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Attachment"
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [ "document-controller" ],
                "summary": "postAttachment",
                "operationId": "postAttachmentUsingPOST",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "in": "body",
                    "name": "params",
                    "description": "params",
                    "required": true,
                    "schema": {
                        "type": "array",
                        "items": {
                            "$ref": "#/definitions/AttachmentCreate"
                        }
                    }
                }, {
                    "name": "file",
                    "in": "query",
                    "description": "file",
                    "required": false,
                    "type": "array",
                    "items": {
                        "type": "file"
                    },
                    "collectionFormat": "multi"
                } ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Attachment"
                        }
                    }
                }
            }
        },
        "/documents/{document_id}/attachments/{attachment_id}": {
            "get": {
                "tags": [ "document-controller" ],
                "summary": "getAttachment",
                "operationId": "getAttachmentUsingGET",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "document_id",
                    "in": "path",
                    "description": "document_id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "attachment_id",
                    "in": "path",
                    "description": "attachment_id",
                    "required": true,
                    "type": "string"
                } ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Attachment"
                        }
                    }
                }
            },
            "put": {
                "tags": [ "document-controller" ],
                "summary": "putAttachment",
                "operationId": "putAttachmentUsingPUT",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "document_id",
                    "in": "path",
                    "description": "document_id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "attachment_id",
                    "in": "path",
                    "description": "attachment_id",
                    "required": true,
                    "type": "string"
                }, {
                    "in": "body",
                    "name": "body",
                    "description": "body",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/AttachmentUpdate"
                    }
                } ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/Attachment"
                        }
                    }
                }
            },
            "delete": {
                "tags": [ "document-controller" ],
                "summary": "deleteAttachment",
                "operationId": "deleteAttachmentUsingDELETE",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "document_id",
                    "in": "path",
                    "description": "document_id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "attachment_id",
                    "in": "path",
                    "description": "attachment_id",
                    "required": true,
                    "type": "string"
                } ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/documents/{document_id}/attachments/{attachment_id}/content": {
            "get": {
                "tags": [ "document-controller" ],
                "summary": "getAttachmentContent",
                "operationId": "getAttachmentContentUsingGET",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "document_id",
                    "in": "path",
                    "description": "document_id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "attachment_id",
                    "in": "path",
                    "description": "attachment_id",
                    "required": true,
                    "type": "string"
                } ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "tags": [ "document-controller" ],
                "summary": "putAttachmentContent",
                "operationId": "putAttachmentContentUsingPUT",
                "consumes": [ "multipart/form-data" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "document_id",
                    "in": "path",
                    "description": "document_id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "attachment_id",
                    "in": "path",
                    "description": "attachment_id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "file",
                    "in": "formData",
                    "description": "file",
                    "required": false,
                    "type": "file"
                } ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/documents/{document_id}/content": {
            "get": {
                "tags": [ "document-controller" ],
                "summary": "getContent",
                "operationId": "getContentUsingGET",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "document_id",
                    "in": "path",
                    "description": "document_id",
                    "required": true,
                    "type": "string"
                } ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/folders/{folder_id}/subfolders/{subfolder_id}/documents/{document_id}/content": {
            "get": {
                "tags": [ "folder-controller" ],
                "summary": "getContent",
                "operationId": "getContentUsingGET_1",
                "consumes": [ "application/json" ],
                "produces": [ "*/*" ],
                "parameters": [ {
                    "name": "folder_id",
                    "in": "path",
                    "description": "folder_id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "subfolder_id",
                    "in": "path",
                    "description": "subfolder_id",
                    "required": true,
                    "type": "string"
                }, {
                    "name": "document_id",
                    "in": "path",
                    "description": "document_id",
                    "required": true,
                    "type": "string"
                } ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Attachment": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mime": {
                    "type": "string"
                },
                "size": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        },
        "AttachmentCreate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
        "AttachmentUpdate": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        },
        "Document": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "folder": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "info": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Entity"
                    }
                },
                "link": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "published_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subfolder": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
        "DocumentUpdate": {
            "type": "object",
            "properties": {
                "info": {
                    "type": "array",
                    "description": "List of infos about the document",
                    "items": {
                        "$ref": "#/definitions/Entity"
                    }
                },
                "name": {
                    "type": "string",
                    "description": "document name"
                },
                "status": {
                    "type": "string",
                    "description": "status of the document"
                },
                "title": {
                    "type": "string",
                    "description": "title of the document"
                }
            }
        },
        "Entity": {
            "type": "object",
            "additionalProperties": {
                "type": "object"
            }
        },
...
    }
}